% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgeneR_class.R
\name{fit_edgeR,fitObject-method}
\alias{fit_edgeR,fitObject-method}
\title{Fit GLM on parent and hybrid crosses using edgeR.}
\usage{
\S4method{fit_edgeR}{fitObject}(object, ...)
}
\arguments{
\item{object}{A \code{fitObject} instance containing counts, metadata, and design.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{fitObject} with updated slots for raw p-values, adjusted FDRs, and model weights.
}
\description{
Fits a negative binomial GLM using edgeR for a given \code{fitObject}, testing for cis and trans effects between homozygous parents and their heterozygous crosses.
}
