% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgeneR_plotting.R
\name{plotPvalHistograms}
\alias{plotPvalHistograms}
\title{Plot p-value histograms nd false discovery rates for certain condition combination.}
\usage{
plotPvalHistograms(fitObject, combo = NULL)
}
\arguments{
\item{fitObject}{A \code{fitObject} that has been fitted using \code{fit_edgeR}.}

\item{combo}{A string specifying the experimental combination (e.g., "Treatment*Time").}
}
\value{
A ggplot object with the p-value histograms.
}
\description{
Plots histograms of raw p-values for a given experimental combination.
}
